# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API E-Commerce Sederhana",
        "description": "Dokumentasi API Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/items": {
            "get": {
                "summary": "Ambil semua item",
                "operationId": "2cbf07eb95d033f8775c36424ac8fa06",
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Item"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Tambah item",
                "operationId": "af37857663e6805dd367a55f374c9ee1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mouse"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Item"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "summary": "Ambil item berdasarkan ID",
                "operationId": "7f1ffbd9fa2b704c61ebcccb636ebb27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Item"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            },
            "put": {
                "summary": "Update seluruh data item",
                "operationId": "4ac9fb55d8a18bd1cba3a29d84f9fd4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            },
            "delete": {
                "summary": "Hapus item",
                "operationId": "0f88b60b73861d5414f8ea2fedd3dfc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            },
            "patch": {
                "summary": "Update sebagian data item",
                "operationId": "8171e1e0bd396fd89495538eb21a8869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Item": {
                "required": [
                    "id",
                    "name",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Laptop"
                    },
                    "price": {
                        "type": "number",
                        "example": 1500000
                    }
                },
                "type": "object"
            }
        }
    }
}